fix: added jsonproperty to dcObject, fixed typo
