+++ server/src/main/kotlin/delta/codecharacter/server/daily_challenge/DailyChallengeEntity.kt
@@ -1,6 +1,5 @@
 package delta.codecharacter.server.daily_challenge
 
-import com.fasterxml.jackson.annotation.JsonProperty
 import delta.codecharacter.dtos.ChallengeTypeDto
 import org.springframework.data.annotation.Id
 import org.springframework.data.mongodb.core.mapping.Document
@@ -9,9 +8,9 @@ import java.util.UUID
 @Document(collection = "daily_challenges")
 data class DailyChallengeEntity(
     @Id val id: UUID,
-    @field:JsonProperty("day") val day: Int,
-    @field:JsonProperty("challName") val challName: String,
-    @field:JsonProperty("challType") val challType: ChallengeTypeDto,
-    @field:JsonProperty("chall") val chall: String,
-    @field:JsonProperty("description") val description: String?,
+    val day: Int,
+    val challName: String,
+    val challType: ChallengeTypeDto,
+    val chall: String,
+    val description: String?,
 )
