--- server/src/main/kotlin/delta/codecharacter/server/daily_challenge/DailyChallengeObject.kt
@@ -1,11 +1,0 @@
-package delta.codecharacter.server.daily_challenge
-
-import delta.codecharacter.dtos.ChallengeTypeDto
-
-data class DailyChallengeObject(
-    val day: Int,
-    val challName: String,
-    val challType: ChallengeTypeDto,
-    val chall: String,
-    val description: String?,
-)
